Add testing framework (jest, supertest)
