+++ package.json
@@ -10,7 +10,8 @@
     "fix": "gts fix",
     "prestart": "npm run-script compile",
     "start": "node .",
-    "watch": "nodemon --ignore build --ext ts --exec 'tsc --incremental && node .'"
+    "watch": "nodemon --ignore build --ext ts --exec 'tsc --incremental && node .'",
+    "test": "jest"
   },
   "dependencies": {
     "@types/passport": "^1.0.7",
@@ -28,10 +29,14 @@
   },
   "devDependencies": {
     "@types/express": "^4.17.13",
+    "@types/jest": "^27.0.2",
     "@types/node": "^16.6.2",
+    "@types/supertest": "^2.0.11",
     "gts": "^3.1.0",
-    "jest": "^27.0.6",
+    "jest": "^27.2.5",
     "nodemon": "^2.0.12",
+    "supertest": "^6.1.6",
+    "ts-jest": "^27.0.5",
     "ts-node": "^10.2.1",
     "typescript": "^4.3.5"
   }
