+++ src/core.ts
@@ -1,0 +1,6 @@
+import express from 'express';
+
+export const app = express();
+
+// Only parse query parameters into strings, not objects
+app.set('query parser', 'simple');
