+++ src/index.test.ts
@@ -1,0 +1,7 @@
+import {app} from './routes';
+import request from "supertest";
+
+test('check is up', async () => {
+    const result = await request(app).get("/up");
+    expect(result.statusCode).toEqual(200);
+});
